pyproject.toml
[tool.poetry]
name = "kak-spell"
version = "0.2.1"
description = "PyEnchant wrapper for Kakoune"
authors = ["Dimitri Merejkowsky <d.merej@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/dmerejkowsky/kak-spell"

[tool.poetry.dependencies]
python = "^3.6"

attrs = "^19.3.0"
pyxdg = "^0.26"
pyenchant = "3.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.3.5"
pytest-mock = "^2.0.0"
mypy = "^0.812"

[tool.poetry.scripts]
kak-spell = "kak_spell.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
